Put newly created fixture in new organization folder
