+++ datalad_registry/tests/conftest.py
@@ -7,10 +7,7 @@ import string
 from subprocess import PIPE, run
 from time import sleep
 
-from datalad import api as dl
-from datalad.api import Dataset
 import pytest
-from pytest import TempPathFactory
 from sqlalchemy.engine.url import URL
 
 from datalad_registry.factory import create_app
@@ -135,55 +132,3 @@ def client(app_instance):
     `app_instance` fixture instead.
     """
     return app_instance.app.test_client()
-
-
-@pytest.fixture(scope="session")
-def empty_ds_annex(tmp_path_factory) -> Dataset:
-    """
-    An empty dataset that is a git-annex repository
-    """
-    return dl.create(path=tmp_path_factory.mktemp("empty_ds_annex"))
-
-
-@pytest.fixture(scope="session")
-def empty_ds_non_annex(tmp_path_factory) -> Dataset:
-    """
-    An empty dataset that is not a git-annex repository
-    """
-    return dl.create(path=tmp_path_factory.mktemp("empty_ds_non_annex"), annex=False)
-
-
-def two_files_ds(annex: bool, tmp_path_factory: TempPathFactory) -> Dataset:
-    """
-    A dataset with two simple files
-    """
-    ds: Dataset = dl.create(
-        path=tmp_path_factory.mktemp(
-            f"two_files_ds_{'annex' if annex else 'non_annex'}"
-        ),
-        annex=annex,
-    )
-
-    file_names = ["file1.txt", "file2.txt"]
-    for file_name in file_names:
-        with open(ds.pathobj / file_name, "w") as f:
-            f.write(f"Hello in {file_name}\n")
-
-    ds.save(message=f"Add {', '.join(file_names)}")
-    return ds
-
-
-@pytest.fixture(scope="session")
-def two_files_ds_annex(tmp_path_factory) -> Dataset:
-    """
-    A dataset with two simple files that is a git-annex repository
-    """
-    return two_files_ds(annex=True, tmp_path_factory=tmp_path_factory)
-
-
-@pytest.fixture(scope="session")
-def two_files_ds_non_annex(tmp_path_factory) -> Dataset:
-    """
-    A dataset with two simple files that is not a git-annex repository
-    """
-    return two_files_ds(annex=False, tmp_path_factory=tmp_path_factory)
