+++ datalad_registry/tests/test_new_organization/conftest.py
@@ -1,6 +1,9 @@
 import os
 
+from datalad import api as dl
+from datalad.api import Dataset
 import pytest
+from pytest import TempPathFactory
 from sqlalchemy.engine.url import URL
 
 from datalad_registry.factory import create_app
@@ -63,3 +66,55 @@ def flask_cli_runner(flask_app):
     The fixture of the test cli runner of flask_app
     """
     return flask_app.test_cli_runner()
+
+
+@pytest.fixture(scope="session")
+def empty_ds_annex(tmp_path_factory) -> Dataset:
+    """
+    An empty dataset that is a git-annex repository
+    """
+    return dl.create(path=tmp_path_factory.mktemp("empty_ds_annex"))
+
+
+@pytest.fixture(scope="session")
+def empty_ds_non_annex(tmp_path_factory) -> Dataset:
+    """
+    An empty dataset that is not a git-annex repository
+    """
+    return dl.create(path=tmp_path_factory.mktemp("empty_ds_non_annex"), annex=False)
+
+
+def two_files_ds(annex: bool, tmp_path_factory: TempPathFactory) -> Dataset:
+    """
+    A dataset with two simple files
+    """
+    ds: Dataset = dl.create(
+        path=tmp_path_factory.mktemp(
+            f"two_files_ds_{'annex' if annex else 'non_annex'}"
+        ),
+        annex=annex,
+    )
+
+    file_names = ["file1.txt", "file2.txt"]
+    for file_name in file_names:
+        with open(ds.pathobj / file_name, "w") as f:
+            f.write(f"Hello in {file_name}\n")
+
+    ds.save(message=f"Add {', '.join(file_names)}")
+    return ds
+
+
+@pytest.fixture(scope="session")
+def two_files_ds_annex(tmp_path_factory) -> Dataset:
+    """
+    A dataset with two simple files that is a git-annex repository
+    """
+    return two_files_ds(annex=True, tmp_path_factory=tmp_path_factory)
+
+
+@pytest.fixture(scope="session")
+def two_files_ds_non_annex(tmp_path_factory) -> Dataset:
+    """
+    A dataset with two simple files that is not a git-annex repository
+    """
+    return two_files_ds(annex=False, tmp_path_factory=tmp_path_factory)
